Test 'nullable list elements'
