+++ generator/test/generator_test_io/nullable_list_elements.log
@@ -1,0 +1,13 @@
+Nullable realm objects are not allowed in collections
+
+in: asset:pkg/test/generator_test_io/nullable_list_elements.dart:14:8
+    ╷
+8   │ @RealmModel()
+9   │ class _Bad {
+    │       ━━━━ in realm model for 'Bad'
+... │
+14  │   late List<_Other?> wrong;
+    │        ^^^^^^^^^^^^^ which has a nullable realm object element type
+    ╵
+Ensure element type is non-nullable
+
